inline behavior into test file
